requirements.txt
numpy==1.24.4
psychopy==2026.1.0
PyAudio==0.2.14
pyserial==3.5
scipy==1.26.4
scikit-learn==1.7.2
joblib==1.5.3
